requirements.txt
aiogram==3.0.0b7
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
